+++ karma.conf.js
@@ -30,6 +30,12 @@ module.exports = (config) => {
     // webpack watches dependencies
     webpack: webpackConfig,
 
+    webpackMiddleware: {
+      stats: {
+        chunks: false,
+      },
+    },
+
     // test results reporter to use
     // possible values: 'dots', 'progress'
     // available reporters: https://npmjs.org/browse/keyword/karma-reporter