+++ packages/ak-layer/src/index.js
@@ -1,23 +1,15 @@
 /** @jsx vdom */
 
-import { vdom, prop } from 'skatejs';
+import { vdom, prop, define } from 'skatejs';
 import { Alignment } from 'akutil-common';
 
-export default {
+const definition = {
   props: {
     position: prop.string({ attribute: true, default: 'right middle' }),
     attachment: prop.string({ attribute: true, default: 'window' }),
     target: { attribute: true },
     renderElementTo: prop.string({ attribute: true }),
     doNotMoveInDOM: prop.boolean({ attribute: true, default: true }),
-    open: prop.string({
-      attribute: true,
-      set(elem) {
-        if (elem.alignment) {
-          elem.alignment.reposition();
-        }
-      },
-    }),
   },
   attached(elem) {
     if (!elem.alignment) {
@@ -42,3 +34,5 @@ export default {
     );
   },
 };
+
+export default define('ak-layer', definition);