+++ packages/akutil-component-template/stories/example-story.js
@@ -1,12 +1,11 @@
 import { storiesOf, action } from '@kadira/storybook';
 import reactify from 'akutil-react';
-import definition from '../src/index';
-import { define } from 'skatejs';
-const { React, ReactDOM, uniqueWebComponent } = window;
+import AkUtilComponentTemplate from '../src/index';
+const { React, ReactDOM } = window;
 import { name } from '../package.json';
-import styles from 'style!./../src/host.less'; // eslint-disable-line import/no-unresolved
+import styles from 'style!./../src/host.less';
 
-const Component = reactify(uniqueWebComponent(name, definition, define), {
+const Component = reactify(AkUtilComponentTemplate, {
   React,
   ReactDOM,
 });
@@ -14,6 +13,9 @@ const Component = reactify(uniqueWebComponent(name, definition, define), {
 storiesOf(name, module)
   .add('a simple akutil-component-template', () => (
     <Component />
+  ))
+  .add('a simple akutil-component-template with a name', () => (
+    <Component name="MyComponent" />
   ))
   .add('an akutil-component-template that emits an action when it is clicked', () => (
     <Component id="myComponent" onClick={action('clicking the WebComponent')} />