+++ packages/akutil-component-template/test/index.js
@@ -1,7 +1,7 @@
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
 import { symbols } from 'skatejs';
-import AkutilComponentTemplate from '../src/index.register.js';
+import AkutilComponentTemplate from '../src/index.js';
 
 chai.use(chaiAsPromised);
 chai.should();
@@ -11,6 +11,6 @@ const expect = chai.expect; // eslint-disable-line no-unused-vars
 describe('akutil-component-template', () => {
   it('should be possible to create a component', () => {
     const component = new AkutilComponentTemplate();
-    component[symbols.shadowRoot].innerHTML.should.match(/I am an .+? element!/);
+    component[symbols.shadowRoot].innerHTML.should.match(/My name is .+?!/);
   });
 });