+++ packages/pf-mention/src/wc/pf-mention-item.js
@@ -1,8 +1,8 @@
-import headStyles from 'style!../host.less'; // eslint-disable-line no-unused-vars, import/no-unresolved, max-len
+import 'style!../host.less';
 import shadowStyles from './pf-mention-item-shadow.less';
 import { localProp } from './skate-local-props';
 
-import akAvatar from 'ak-avatar'; // eslint-disable-line no-unused-vars
+import 'ak-avatar';
 import { define, vdom, prop, emit, state } from 'skatejs'; // eslint-disable-line no-unused-vars