+++ packages/pf-mention/src/wc/pf-mention-list.js
@@ -1,4 +1,4 @@
-import headStyles from 'style!../host.less'; // eslint-disable-line no-unused-vars, import/no-unresolved, max-len
+import 'style!../host.less';
 import shadowStyles from './pf-mention-list-shadow.less';
 import mentionItem from './pf-mention-item'; // eslint-disable-line no-unused-vars
 import scrollable from './pf-scrollable'; // eslint-disable-line no-unused-vars
@@ -167,14 +167,12 @@ const definition = {
       <div>
         <style>{shadowStyles.toString()}</style>
         <div className={classes}>
-          <div className={styles.innerList}>
-            <pf-scrollable
-              className={styles.scrollable}
-              ref={(ref) => { elem._scrollable = ref; }}
-            >
-              {renderItems(elem)}
-            </pf-scrollable>
-          </div>
+          <pf-scrollable
+            className={styles.scrollable}
+            ref={(ref) => { elem._scrollable = ref; }}
+          >
+            {renderItems(elem)}
+          </pf-scrollable>
         </div>
       </div>
     );